refactor data_structures/location
